Address string concatenation comments
