+++ .github/workflows/go.yml
@@ -34,4 +34,4 @@ jobs:
       run: go build -v .
 
     - name: Test
-      run: go test ./test -v .
+      run: go test github.com/googleinterns/ocsp-response-linter/test -v .
